+++ src/pages/IndividualFeed.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import backgroundImage from "../assets/images/IndividualFeed-BackgroundImage.png";
 import logo from "../assets/images/logo.png";
 import facebook from "../assets/images/Facebook.png";
@@ -20,6 +20,7 @@ import { useSubjectInfo } from "../hooks/useSubjectInfo";
 import { useIndividualQuestions } from "../hooks/useIndividualQuestions";
 import FeedCardPlaceholder from "../components/FeedCardPlaceholder";
 import { RotatingAnimation } from "../styles/rotatingAnimation";
+import { useScroll } from "../hooks/useScroll";
 
 const IndividualFeed = () => {
   const [offset, setOffset] = useState(0);
@@ -29,21 +30,8 @@ const IndividualFeed = () => {
     offset,
     limit: LIMIT,
   });
+  const { moreData } = useScroll({ setOffset, questionInfo, LIMIT });
   const [loading, setLoading] = useState(true);
-  const [moreData, setMoreData] = useState(false);
-
-  // 무한 스크롤
-  const handleScroll = useCallback(() => {
-    if (
-      window.innerHeight + window.scrollY >=
-      document.body.offsetHeight - 50
-    ) {
-      if (moreData === false) {
-        setMoreData(true);
-        setOffset((prev) => prev + LIMIT);
-      }
-    }
-  }, [moreData]);
 
   // 스켈리톤 ui를 위한 상태 변경
   useEffect(() => {
@@ -53,19 +41,6 @@ const IndividualFeed = () => {
       setLoading(true);
     }
   }, [userInfo, questionInfo]);
-
-  // 무한 스크롤 이벤트 등록, 삭제
-  useEffect(() => {
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, [handleScroll]);
-
-  // 추가로 정보를 더 받아와서 배열이 변경되면 moreData를 false로 변경
-  useEffect(() => {
-    if (questionInfo?.length > 0) {
-      setMoreData(false);
-    }
-  }, [questionInfo]);
 
   return (
     <Wrapper>
@@ -123,7 +98,9 @@ const IndividualFeed = () => {
         )}
       </BodyWrapper>
 
-      {(moreData && (questionInfo.length < questionInfo.count)) && <RotatingAnimation />}
+      {moreData && questionInfo.length < questionInfo.count && (
+        <RotatingAnimation />
+      )}
 
       {/* 질문 작성하기 버튼 */}
       <AddQuestion />
